feat: extract shlagemon list component
